requirements.txt
accelerate==1.1.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
diffusers==0.31.0
ecdsa==0.19.0
fastapi==0.115.4
filelock==3.16.1
firebase-admin==6.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jws==0.1.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.1.3
oauth2client==4.1.3
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-jwt==2.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.4
rsa==4.9
safetensors==0.4.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.32.0
watchdog==5.0.3
zipp==3.20.2